CMake set up and made sense of, first c++ homework 75% done, UTF-8 coding tomorrow

+++ School/C++/Homeworks/Hw01/tasks.h
@@ -3,6 +3,7 @@
 #include <utility>
 #include <stdio.h>
 #include <vector>
+#include <limits>
 
 namespace math
 {
@@ -13,7 +14,7 @@ namespace strutils
 {
     std::string camel_to_snake(std::string input);
 
-    bool parse_uint(std::string input, int &result);
+    bool parse_uint(std::string input, uint32_t &result);
 
-    bool validate_utf8(std::vector<size_t> input, size_t &result);
+    bool validate_utf8(std::vector<uint8_t> input, size_t &result);
 }
